moved 'assets' files to a service
