+++ src/db/__test__/mocks/adapter.js
@@ -5,7 +5,6 @@ const createSqlProxy = require('./sql');
 const createMockAdapter = (driver, sql = createSqlProxy()) =>
   createAdapter({
     taskedDbDriver: driver,
-    batchQueryFn: () => x => x,
     errorFactory: () => x => x,
     sql,
   });
