+++ src/db/index.js
@@ -3,7 +3,6 @@ const createAdapter = require('./adapter');
 
 // adapter dependencies
 const { toDbError } = require('../errorHandling');
-const { batchQuery } = require('./utils');
 const sql = require('./sql');
 
 module.exports = {
@@ -13,7 +12,6 @@ module.exports = {
     createAdapter({
       taskedDbDriver: driver,
       errorFactory: toDbError,
-      batchQueryFn: batchQuery,
       sql,
     }),
 };
