+++ src/db/utils/index.js
@@ -4,12 +4,16 @@ const pgp = require('../driver/pgp');
 const addParens = s => `(${s})`;
 
 const formatPairs = R.pipe(
-  R.map(R.pipe(pgp.as.csv, addParens)),
+  R.map(
+    R.pipe(
+      pgp.as.csv,
+      addParens
+    )
+  ),
   R.join(','),
   addParens
 );
 
 module.exports = {
   formatPairs,
-  batchQuery: require('./batchQuery'),
 };
