+++ src/services/assets/resolver/index.js
@@ -1,0 +1,21 @@
+const { identity } = require('ramda');
+const create = require('../../../resolvers/create');
+
+const oneConfig = {
+  ...require('./validation/one'),
+  transformResult: require('./transformResult/one'),
+  dbQuery: identity,
+};
+
+const manyConfig = {
+  ...require('./validation/many'),
+  transformResult: require('./transformResult/many'),
+  dbQuery: identity,
+};
+
+module.exports = {
+  one: ({ getData, emitEvent }) =>
+    create.one(oneConfig)({ db: getData, emitEvent }),
+  many: ({ getData, emitEvent }) =>
+    create.many(manyConfig)({ db: getData, emitEvent }),
+};
