+++ src/services/assets/resolver/transformResult/__test__/many.test.js
@@ -3,7 +3,7 @@ const Maybe = require('folktale/maybe');
 const { Asset, List } = require('../../../../types');
 const transformResult = require('../many');
 
-const { asset, transformedAsset } = require('../../mocks/asset');
+const { asset, transformedAsset } = require('./mocks');
 
 const assetsFromDb = [
   Maybe.of({
