+++ src/services/assets/resolver/transformResult/__test__/mocks.js
