+++ src/services/assets/resolver/transformResult/common.js
