+++ src/services/assets/resolver/transformResult/many.js
