+++ src/services/assets/resolver/validation/one.js
